+++ src/main/java/net/pedroksl/advanced_ae/client/renderer/QuantumArmorRenderer.java
@@ -14,8 +14,6 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EquipmentSlot;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.api.distmarker.OnlyIn;
 import net.pedroksl.advanced_ae.AdvancedAE;
 import net.pedroksl.advanced_ae.common.items.armors.QuantumArmorBase;
 import net.pedroksl.ae2addonlib.util.Colors;
@@ -63,7 +61,6 @@ public class QuantumArmorRenderer extends GeoArmorRenderer<QuantumArmorBase> {
         super.applyBoneVisibilityBySlot(currentSlot);
     }
 
-    @OnlyIn(Dist.CLIENT)
     private void setVisibleBySlot(EquipmentSlot slot) {
         HumanoidModel<?> model = this;