+++ src/main/java/net/pedroksl/advanced_ae/events/AAEPlayerEvents.java
@@ -1,6 +1,5 @@
 package net.pedroksl.advanced_ae.events;
 
-import net.minecraft.client.Minecraft;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.tags.FluidTags;
 import net.minecraft.world.effect.MobEffectInstance;
@@ -14,7 +13,6 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.phys.Vec3;
 import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.neoforge.client.event.RenderPlayerEvent;
 import net.neoforged.neoforge.common.NeoForgeMod;
 import net.neoforged.neoforge.event.ItemAttributeModifierEvent;
 import net.neoforged.neoforge.event.entity.player.PlayerEvent;
@@ -25,7 +23,6 @@ import net.pedroksl.advanced_ae.common.definitions.AAEComponents;
 import net.pedroksl.advanced_ae.common.items.armors.*;
 import net.pedroksl.advanced_ae.common.items.upgrades.UpgradeType;
 import net.pedroksl.advanced_ae.network.packet.ItemTrackingPacket;
-import net.pedroksl.advanced_ae.network.packet.KeysPressedPacket;
 import net.pedroksl.advanced_ae.xmod.Addons;
 import net.pedroksl.advanced_ae.xmod.apoth.ApoEnchPlugin;
 
@@ -38,7 +35,7 @@ public class AAEPlayerEvents {
     public static final String DOWN_KEY_DATA = "aae$downkey";
 
     @SubscribeEvent
-    public static void ItemAttributes(ItemAttributeModifierEvent event) {
+    public static void itemAttributes(ItemAttributeModifierEvent event) {
         ItemStack itemStack = event.getItemStack();
         if (itemStack.getItem() instanceof QuantumArmorBase armor) {
             if (armor.isUpgradeEnabledAndPowered(itemStack, UpgradeType.STEP_ASSIST)) {
@@ -74,7 +71,7 @@ public class AAEPlayerEvents {
 
     @SuppressWarnings("deprecation")
     @SubscribeEvent
-    public static void BreakSpeed(PlayerEvent.BreakSpeed event) {
+    public static void breakSpeed(PlayerEvent.BreakSpeed event) {
         Player player = event.getEntity();
         if (!player.onGround()) {
             if (!Addons.APOTHIC_ENCHANTING.isLoaded()
@@ -98,74 +95,6 @@ public class AAEPlayerEvents {
             }
         }
     }
-
-    @SubscribeEvent
-    public static void playerRender(RenderPlayerEvent.Pre event) {
-        Player player = event.getEntity();
-        if (!(player instanceof ServerPlayer)) {
-            var renderer = event.getRenderer();
-            var model = renderer.getModel();
-
-            ItemStack helmetStack = player.getItemBySlot(EquipmentSlot.HEAD);
-            if (helmetStack.getItem() instanceof QuantumHelmet item && item.isVisible(helmetStack)) {
-                model.hat.visible = false;
-            }
-
-            ItemStack chestStack = player.getItemBySlot(EquipmentSlot.CHEST);
-            if (chestStack.getItem() instanceof QuantumChestplate item && item.isVisible(chestStack)) {
-                model.leftSleeve.visible = false;
-                model.rightSleeve.visible = false;
-                model.jacket.visible = false;
-
-                model.leftArm.visible = false;
-                model.rightArm.visible = false;
-            }
-
-            ItemStack leggingsStack = player.getItemBySlot(EquipmentSlot.LEGS);
-            ItemStack bootsStack = player.getItemBySlot(EquipmentSlot.FEET);
-            if (leggingsStack.getItem() instanceof QuantumLeggings item && item.isVisible(leggingsStack)) {
-                model.leftPants.visible = false;
-                model.rightPants.visible = false;
-
-                if (bootsStack.getItem() instanceof QuantumBoots item2 && item2.isVisible(bootsStack)) {
-                    model.leftLeg.visible = false;
-                    model.rightLeg.visible = false;
-                }
-            }
-        }
-    }
-
-    //    @SubscribeEvent
-    //    public static void playerArmRender(RenderArmEvent event) {
-    //        Player player = event.getPlayer();
-    //
-    //        ItemStack stack = player.getItemBySlot(EquipmentSlot.CHEST);
-    //        if (stack.getItem() instanceof QuantumChestplate chest) {
-    //            var renderProvider = chest.getAnimatableInstanceCache().getRenderProvider();
-    //            if (renderProvider instanceof GeoRenderProvider provider) {
-    //                var renderer = provider.getGeoArmorRenderer(player, stack, EquipmentSlot.CHEST, null);
-    //                if (renderer instanceof QuantumArmorRenderer quantumRenderer) {
-    //                    var boneName = event.getArm() == HumanoidArm.RIGHT ? QuantumArmorRenderer.RIGHT_ARM :
-    //                            QuantumArmorRenderer.LEFT_ARM;
-    //                    var model = quantumRenderer.getGeoModel();
-    //                    var renderType = RenderType.armorCutoutNoCull(model.getTextureResource(chest));
-    //
-    //                    var poseStack = event.getPoseStack();
-    //                    poseStack.pushPose();
-    //                    poseStack.last().pose().rotate((float) Math.PI,0, 0, 1);
-    //                    poseStack.translate(-1, 0.5f, 0);
-    //                    poseStack.translate(-0.3125F, -0.125F, 0);
-    //                    var source = event.getMultiBufferSource();
-    //                    var buffer = source.getBuffer(renderType);
-    //                    quantumRenderer.renderChildBones(boneName, poseStack, chest, renderType,
-    //                            source, buffer, false, 0, event.getPackedLight(), OverlayTexture.NO_OVERLAY, 0);
-    //                    poseStack.popPose();
-    //
-    //                    event.setCanceled(true);
-    //                }
-    //            }
-    //        }
-    //    }
 
     @SubscribeEvent
     public static void playerTick(PlayerTickEvent.Pre event) {
@@ -219,46 +148,6 @@ public class AAEPlayerEvents {
                             / 25f;
                     var direction = upKey ? 1 : -1;
                     player.moveRelative(value, new Vec3(0, direction, 0));
-                }
-            }
-        }
-    }
-
-    @SubscribeEvent
-    public static void playerTick(PlayerTickEvent.Post event) {
-        Player player = event.getEntity();
-        if (!(player instanceof ServerPlayer)) {
-            ItemStack bootStack = player.getItemBySlot(EquipmentSlot.FEET);
-            if (bootStack.getItem() instanceof QuantumBoots boots) {
-                if (boots.isUpgradeEnabledAndPowered(bootStack, UpgradeType.FLIGHT_DRIFT)) {
-                    var options = Minecraft.getInstance().options;
-                    var noKey = !options.keyUp.isDown()
-                            && !options.keyRight.isDown()
-                            && !options.keyDown.isDown()
-                            && !options.keyLeft.isDown();
-                    if (player.getPersistentData().getBoolean(NO_KEY_DATA) != noKey) {
-                        // Send packet to server if data on player is different
-                        PacketDistributor.sendToServer(new KeysPressedPacket(NO_KEY_DATA, noKey));
-                        player.getPersistentData().putBoolean(NO_KEY_DATA, noKey);
-                    }
-                }
-            }
-
-            ItemStack chestStack = player.getItemBySlot(EquipmentSlot.CHEST);
-            if (chestStack.getItem() instanceof QuantumChestplate) {
-                var options = Minecraft.getInstance().options;
-                var downKey = options.keyShift.isDown();
-                if (player.getPersistentData().getBoolean(DOWN_KEY_DATA) != downKey) {
-                    // Send packet to server if data on player is different
-                    PacketDistributor.sendToServer(new KeysPressedPacket(DOWN_KEY_DATA, downKey));
-                    player.getPersistentData().putBoolean(DOWN_KEY_DATA, downKey);
-                }
-
-                var upKey = options.keyJump.isDown();
-                if (player.getPersistentData().getBoolean(UP_KEY_DATA) != upKey) {
-                    // Send packet to server if data on player is different
-                    PacketDistributor.sendToServer(new KeysPressedPacket(UP_KEY_DATA, upKey));
-                    player.getPersistentData().putBoolean(UP_KEY_DATA, upKey);
                 }
             }
         }