+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-## Bug Fixes
-- Fixed the reaction chamber getting stuck if inputs are split in multiple slots
+- Added extra client side checks for player events